requirements.txt
flask
pypdf2
langchain
langchain-google-genai
google-generativeai
python-dotenv